Learning rate schedules, ignore padding in loss and add smoke tests (#20)
The main motivation for this was to be able to use the linear decay with warmup LR schedule used in most of the BERT fine-tuning litterature. It gives the ability to chose a LR decay shape: exponential with a 0.95 decay rate (default and the previous behaviour), linear or constant, and a number of linear warmup steps (defaulting to 0). Along the way I noticed that the CrossEntropyLoss was taking into account the performances on the padding tokens (which I removed) and I finally added some smoke test that can be run either in local or in github CI and which ensures that we are at least able to learn and parse a toy treebank and eval the results without crashing.

Overall, this has no groundbreaking impact on scores. It does make the parser learn really faster (in terms of epochs) and reduce the computation and memory load.

## Added

- A new config option `lr_schedule` of the form
  ```yaml
  lr_schedule:
    shape: linear
    warmup_steps: 100
  ```
  which allows to chose a decay shape and a number of linear warmup steps
- Examples of training configurations (in `examples/`)
- Smoke tests using tox and used as checks in Github CI to ensure that `graph_parser` at least runs without crashing and outputs a valid model and valid parses. They will also auto-release to Pypi when we create new releases on Github if we set up a Pypi project and add the credentials as repository secrets.

## Changed

- Padding tokens are now ignored in the training and development losses