Allow skipping sentences that transformers models can't encode (#46)
This is meant to answer point (i) of #40 “BERT is internally limited to process 512 BPE max per sentence. Otherwise crash.”. This PR introduces a mechanism to allow skipping sentences longer than that. So far it is only available for `parse` (not in training mode) and with `--raw`. Supporting it for tagged input could be done but would induce more complexity in dataset handling.

## Added

- A `strict` argument to `graph_parser.parse`, defaulting to `False`, that when set to `True` will silently skip sentences that can't be managed by the lexers instead of raising an exception (in raw mode only)
- A `--ignore-unencodable` flag to `hopsparser parse` that activates that behaviour.
- A `--batch-size` option to `hopsparser parse` to chose a batch size different from the training one